refactor: Create exchange-rate folder
