+++ src/app/exchange-rate/exchange-rate-chart.component.scss
