+++ src/app/interceptors/exchange-rate-mock.interceptor.ts
@@ -1,8 +1,9 @@
-import { ExchangeRate, iso4217 } from '../models';
 import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { map, tap } from 'rxjs/operators';
 import { Observable, of } from 'rxjs';
+import { iso4217 } from 'src/app/exchange-rate/iso4217.enum';
+import { ExchangeRate } from '../exchange-rate/exchange-rate.type';
 
 @Injectable()
 export class ExchangeRateMockInterceptor implements HttpInterceptor {
