--- src/app/models/index.ts
@@ -1,3 +1,0 @@
-export * from './exchange-rate.type';
-export * from './exchange-rate-infos.type';
-export * from './iso4217.enum';
