+++ src/app/services/exchange-rate.service.spec.ts
@@ -2,9 +2,9 @@ import { ExchangeRateMockInterceptor } from 'src/app/interceptors';
 import { ExchangeRateService } from './exchange-rate.service';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { iso4217 } from 'src/app/models';
 import { map } from 'rxjs';
 import { TestBed } from '@angular/core/testing';
+import { iso4217 } from 'src/app/exchange-rate/iso4217.enum';
 
 describe('ExchangeRateService', () => {
     let exchangeRateService: ExchangeRateService;
